+++ src/features/reader/viewer/ReaderChapterViewer.tsx
@@ -14,6 +14,7 @@ import Box from '@mui/material/Box';
 import { useLingui } from '@lingui/react/macro';
 import type {
     IReaderSettings,
+    ReaderPageBackgroundColor,
     ReaderPagerProps,
     ReaderPageSpreadState,
     ReaderResumeMode,
@@ -124,6 +125,9 @@ const BaseReaderChapterViewer = ({
     const [pagesToSpreadState, setPagesToSpreadState] = useState<ReaderPageSpreadState[]>(
         READER_DEFAULT_PAGES_STATE.pageSpreadStates,
     );
+    const [pageBackgroundColors, setPageBackgroundColors] = useState<ReaderPageBackgroundColor[]>(
+        READER_DEFAULT_PAGES_STATE.pageBackgroundColors,
+    );
 
     const ref = useRef<HTMLDivElement>(null);
     const isCurrentChapterRef = useRef(isCurrentChapter);
@@ -197,6 +201,13 @@ const BaseReaderChapterViewer = ({
 
                     setPageLoadStates(value);
                 },
+                (value) => {
+                    if (isCurrentChapterRef.current) {
+                        getReaderPagesStore().setPageBackgroundColor(value);
+                    }
+
+                    setPageBackgroundColors(value);
+                },
                 readingMode,
             ),
         [actualPages, readingMode],
@@ -251,6 +262,7 @@ const BaseReaderChapterViewer = ({
         actualPages,
         pageLoadStates,
         pagesToSpreadState,
+        pageBackgroundColors,
         arePagesFetched,
         setArePagesFetched,
         (value) => updateState(value, noOp, getReaderChaptersStore().setReaderStateChapters),
@@ -264,6 +276,12 @@ const BaseReaderChapterViewer = ({
                 value,
                 setPagesToSpreadState,
                 getReaderPagesStore().setPageSpreadStates.bind(getReaderPagesStore()),
+            ),
+        (value) =>
+            updateState(
+                value,
+                setPageBackgroundColors,
+                getReaderPagesStore().setPageBackgroundColor.bind(getReaderPagesStore()),
             ),
         (value) => updateState(value, noOp, getReaderPagesStore().setCurrentPageIndex.bind(getReaderPagesStore())),
         (value) => {